Module Asak.Lambda_utils

type threshold =

| Percent of int
A percentage threshold, a number between 0 and 100.

| Hard of int
A hard threshold.

type config = {

should_sort : bool;
If we sort all lists of hash.

hash_var : bool;
If we hash names in the AST.

}

val hash_lambda : config -> threshold -> Lambda.lambda -> (int * string) * (int * string) list

Hash a lambda expression. Usage: hash_lambda config threshold expr. Two lambda expressions "of the same shape" will share the same hash.

parameter config

A configuration for the function.

parameter threshold

Used to decide which sub-AST hash is kept.

parameter expr

The expression.

returns

A tuple with the main hash, and a list of hashes of sub-ASTs. All hashes are given with the weight of their AST.

val hash_all : config -> int -> ('a * Lambda.lambda) list -> ('a * ((int * string) * (int * string) list)) list

Using a hard threshold, hash a list of lambda expressions from Parse_structure.read_structure.

val escape_hash_list : ('a * ((int * string) * (int * string) list)) list -> ('a * ((int * string) * (int * string) list)) list

Escape hashs.

val inline_all : Lambda.lambda -> Lambda.lambda

Inline all possible (ie. without side effect) expressions in a lambda expression.